test: move simplify tests to unit tests for dc_simplify.rs
